requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
defusedxml==0.7.1
Django==5.1.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-nested-routers==0.94.2
flake8==7.1.1
idna==3.10
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
py==1.11.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pyparsing==3.2.1
pytest==8.3.3
pytest-django==4.9.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
toml==0.10.2
uritemplate==4.1.1
urllib3==2.3.0
